dataset/pyproject.toml
[project]
name = "dataset"
requires-python = ">=3.9,<3.14"
version = "0.1.0"
dependencies = [
    "bs4>=0.0.2",
    "dagster==1.11.13",
    "dagster-gcp>=0.27.13",
    "dagster-gcp-pandas>=0.27.13",
    "duckdb>=1.4.1",
    "google-cloud-bigquery>=3.38.0",
    "httpx>=0.28.1",
    "polars>=1.34.0",
]

[dependency-groups]
dev = [
    "dagster-webserver",
    "dagster-dg-cli",
]

[tool.pyright]
typeCheckingMode = "basic"

[tool.ruff]
line-length = 120

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
force-include = { "pyproject.toml" = "pyproject.toml" }

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "dataset"
registry_modules = [
    "dataset.components.*",
]


